Refine MyBatis conditions and split tests
Tightened conditional annotations by removing unnecessary `@Inherited` usage and making MyBatis availability require both `SqlSession` and `SqlSessionFactory`. Added a dedicated Spring Boot integration test to verify real bean wiring, and refactored the existing auto-configuration test to extend `AutoConfigurationTest` with explicit disabled-property and missing-class conditions.
